feat: parse cli arguments and print them to a file
Essentially the initial commit for this repository. Using this
project in order to learn Rust as it seems like a fun language.
Ultimately, the program will take a name and a date, then enter
those values into a database (most likely just a json file) so
that when requested the user can pull any date referring to
any person previously entered. Furthermore, development of a TUI
calendar for the purpose of better input as well as interactivity
is to be slated.